# api-contract.yaml
openapi: 3.0.3
info:
  title: GroceryHack API
  version: 1.0.0
  description: |
    API contract for the GroceryHack application.
    All endpoints require Authorization header with Bearer token unless marked as public.
    All timestamps are ISO 8601 UTC. All monetary values are decimal numbers.
servers:
  - url: /api/v1

# ────────────────────────────────────────────────────────────
# TAGS
# ────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, postal_code]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                display_name:
                  type: string
                postal_code:
                  type: string
                budget:
                  type: number
                dietary_restrictions:
                  type: array
                  items:
                    type: string
                max_stores:
                  type: integer
                  enum: [1, 2]
                  default: 1
                household_size:
                  type: integer
                  default: 1
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refresh_token:
                    type: string
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refresh_token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      description: |
        Always returns 200 regardless of whether the email exists.
        This prevents email enumeration attacks. If the email exists,
        a reset link is sent. Token expires after 1 hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: If the email exists, a reset link has been sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent:
                    type: boolean
                    description: Always true — never reveal whether the email exists
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a token from the reset email
      security: []
      description: |
        Consumes the reset token and sets the new password.
        Token can only be used once. Expired or used tokens return RESET_TOKEN_INVALID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  reset:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'

  # ══════════════════════════════════════
  # USERS
  # ══════════════════════════════════════
...
  /important-items:
    get:
      tags: [ImportantItems]
      summary: Get all important items for the current user
      parameters:
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
          description: If true, only return active items. If false, return all items including deactivated.
      responses:
        '200':
          description: User's important items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImportantItem'

    post:
      tags: [ImportantItems]
      summary: Add a new important item
      description: If an item with the same name already exists for this user, it is reactivated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                quantity:
                  type: string
      responses:
        '201':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportantItem'
  /important-items/{item_id}:
    patch:
      tags: [ImportantItems]
      summary: Update an important item (name, quantity, or active status)
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                quantity:
                  type: string
                is_active:
                  type: boolean
                  description: Set to false to deactivate (item is preserved for history)
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportantItem'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════════════
  # OPTIMIZE
  # ══════════════════════════════════════
...
  /events:
    post:
      tags: [Events]
      summary: Track analytics events (single or batch)
      description: |
        Accepts a single event or a batch of events. user_id is attached
        from the JWT token. Events are inserted asynchronously — never
        blocks the user experience. Rate limit: 100 events per user per minute.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [event_type]
                  properties:
                    event_type:
                      type: string
                    metadata:
                      type: object
                    session_id:
                      type: string
                    created_at:
                      type: string
                      format: date-time
                - type: object
                  required: [events]
                  properties:
                    events:
                      type: array
                      items:
                        type: object
                        required: [event_type]
                        properties:
                          event_type:
                            type: string
                          metadata:
                            type: object
                          session_id:
                            type: string
                          created_at:
                            type: string
                            format: date-time
      responses:
        '200':
          description: Events received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /events/public:
    post:
      tags: [Events]
      summary: Track public events (no auth required)
      security: []
      description: |
        Only accepts share_link_opened, share_recipient_signed_up, and shared_plan_viewed events.
        Requires a valid plan token or share token in metadata for verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type, metadata]
              properties:
                event_type:
                  type: string
                  enum: [share_link_opened, share_recipient_signed_up, shared_plan_viewed]
                metadata:
                  type: object
      responses:
        '200':
          description: Event received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /events/pixel:
    get:
      tags: [Events]
      summary: Email open tracking pixel
      security: []
      description: |
        Returns a 1x1 transparent GIF. Records an email_opened event.
        Deduplicates — only records first open per plan_token per user.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: user
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 1x1 transparent GIF
          content:
            image/gif:
              schema:
                type: string
                format: binary
  /r:
    get:
      tags: [Events]
      summary: Email click tracking redirect
      security: []
      description: |
        Records an email_clicked event with the link target, then redirects
        to the actual URL. Deduplicates first click only.
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: user
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '302':
          description: Redirect to target URL

  # ══════════════════════════════════════
  # ADMIN
  # ══════════════════════════════════════
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ══════════════════════════════════════
  # SHARED RESPONSES
  # ══════════════════════════════════════
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ══════════════════════════════════════
  # SCHEMAS
  # ══════════════════════════════════════
  schemas:
...
    Error:
      type: object
      required: [error, code, message]
      properties:
        error:
          type: boolean
          enum: [true]
        code:
          type: string
          description: Machine-readable error code
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable message
          example: No stores found near that postal code
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        postal_code:
          type: string
        lat:
          type: number
        lng:
          type: number
        budget:
          type: number
        dietary_restrictions:
          type: array
          items:
            type: string
        max_stores:
          type: integer
        household_size:
          type: integer
        household_members:
          type: array
          items:
            $ref: '#/components/schemas/HouseholdMember'
        household_names:
          type: array
          items:
            type: string
        taste_profile:
          type: object
          additionalProperties:
            type: number
        subscription_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HouseholdMember:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        dietary_restrictions:
          type: array
          items:
            type: string
...
    ImportantItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        deactivated_at:
          type: string
          format: date-time
          nullable: true
...
security:
  - BearerAuth: []